requirements.txt
Django==3.1.7
djangorestframework==3.12.2
mysqlclient==2.0.3
django-mysql==3.11.1
flake8
PyJWT==2.0.1
django-cors-headers==3.7.0
coverage
Pillow
mailjet-rest==1.3.3
dnspython==2.2.1
django-debug-toolbar